pyproject.toml
[tool.poetry]
name = "harlequin-risingwave"
version = "0.1.0"
description = "A Harlequin adapter for Risingwave."
authors = ["ZhengYu, Xu <zen-xu@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
harlequin = "^1.7"
psycopg2-binary = "^2.9.9"
harlequin-postgres = "0.2.2"

[tool.poetry.plugins."harlequin.adapter"]
risingwave = "harlequin_risingwave:HarlequinRisingwaveAdapter"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["A", "B", "E", "F", "I"]
target-version = "py39"

[tool.mypy]
python_version = "3.9"
show_column_numbers = true

# show error messages from unrelated files
follow_imports = "normal"

# be strict
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
strict_optional = true

warn_return_any = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true

no_implicit_reexport = true
strict_equality = true
